+++ rust/main/agents/relayer/src/msg/message_processor/tests.rs
@@ -1,3 +1,1 @@
-mod tests_common;
-mod tests_filter_operations_for_preparation;
-mod tests_operation_disposition_by_payload_status;
+pub(crate) mod tests_common;