+++ typescript/http-registry-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/http-registry-server
 
+## 19.9.0
+
+### Patch Changes
+
+- Updated dependencies [8c027d852]
+- Updated dependencies [8c027d852]
+  - @hyperlane-xyz/utils@19.9.0
+  - @hyperlane-xyz/sdk@19.9.0
+
 ## 19.8.0
 
 ### Patch Changes