+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hyperlane-xyz/radix-sdk
 
+## 19.9.0
+
+### Patch Changes
+
+- Updated dependencies [8c027d852]
+  - @hyperlane-xyz/utils@19.9.0
+
 ## 19.8.0
 
 ### Minor Changes